Renovate Dimension object (#2953)
* add some machine-readable docstrings

* add deprecation msg to assignment operator

* use size_t when dealing with the indexes

* remove some empty lines

* remove unnecessary static_cast

* update pybind export for the constructors

* add more docstrings, add deprecation msgs to cast-to-pointer operators

* more docstrings and a const

* add ingroup docstring

* revert incorrect pybind change, add default argument indicators to dimension.cc, weaken deprecation messages

* fix docstring, change int n() to size_t n()

* use size_t in for loops

* delay deprecation notices by one release

* adopt std::copy suggestion


* Remove superfluous const qualifiers from ints (suggested by Evangelista)

* try fixing the std::copy suggestion

* touch up export_mints.cc

forgot to remove the constructor const qualifiers after updating the dimension .h and .cc

---------

+++ psi4/src/export_mints.cc
@@ -319,8 +319,8 @@ void export_mints(py::module& m) {
     typedef void (Vector::*vector_three)(double alpha, double beta, const Vector &other);
 
     py::class_<Dimension>(m, "Dimension", "Initializes and defines Dimension Objects")
-        .def(py::init<int>())
-        .def(py::init<int, const std::string&>())
+        .def(py::init<size_t>())
+        .def(py::init<size_t, const std::string&>())
         .def(py::init<const std::vector<int>&>())
         .def("print_out", &Dimension::print, "Print out the dimension object to the output file")
         .def("init", &Dimension::init, "Re-initializes the dimension object")
